* Removed old heat load fill_dict * added options to force output or subset of fills * changed structure!! * added integrated heat load

+++ 025_create_large_hl_dict.py
@@ -21,23 +21,32 @@ import GasFlowHLCalculator.qbs_fill as qf
 
 # Config
 subtract_offset = True
+average_offset_seconds = 600
 hrs_after_sb = 24
 hl_dict_dir = './hl_dicts/'
 
 # Arguments
 parser = argparse.ArgumentParser()
 parser.add_argument('year', type=int)
+parser.add_argument('-o', help='Force output filename', type=str)
+parser.add_argument('--fills', help='Force fill list', nargs='+')
 args = parser.parse_args()
 
-# Find new version
-re_version = re.compile('large_heat_load_dict_%i_(\d+).pkl' % args.year)
-matches = filter(None, map(re_version.match, os.listdir(hl_dict_dir)))
-versions = map(lambda x: int(x.group(1)), matches)
-version = max(versions)+1
+if args.fills and not args.o:
+    raise ValueError('If fills are specified, thet output has to, as well!')
 
-# Config again
-pkl_file_name = hl_dict_dir + 'large_heat_load_dict_%i_%i.pkl' % (args.year, version)
-latest_pkl = hl_dict_dir + 'large_heat_load_dict_%i_latest.pkl' % args.year
+# Find new version
+if args.o:
+    pkl_file_name = args.o
+else:
+    re_version = re.compile('^large_heat_load_dict_%i_(\d+).pkl$' % args.year)
+    matches = filter(None, map(re_version.match, os.listdir(hl_dict_dir)))
+    versions = map(lambda x: int(x.group(1)), matches)
+    version = max(versions)+1
+
+    pkl_file_name = hl_dict_dir + 'large_heat_load_dict_%i_%i.pkl' % (args.year, version)
+    latest_pkl = hl_dict_dir + 'large_heat_load_dict_%i_latest.pkl' % args.year
+
 logfile = pkl_file_name + '.log'
 
 if args.year == 2016:
@@ -46,14 +55,12 @@ if args.year == 2016:
     fills_bmodes_file = base_folder + '/fills_and_bmodes.pkl'
     csv_file_names = ['fill_basic_data_csvs/basic_data_fill_%d.csv',
             'fill_bunchbybunch_data_csvs/bunchbybunch_data_fill_%d.csv']
-    h5_file_names = ['heatloads_fill_h5s/heatloads_all_fill_%i.h5']
     filling_pattern_csv = base_folder + './fill_basic_data_csvs/injection_scheme.csv'
 elif args.year == 2015:
     base_folder = '/afs/cern.ch/project/spsecloud/'
     child_folders = ['LHC_2015_PhysicsAfterTS2/', 'LHC_2015_PhysicsAfterTS3/', 'LHC_2015_Scrubbing50ns/', 'LHC_2015_IntRamp50ns/', 'LHC_2015_IntRamp25ns/']
     fills_bmodes_file = base_folder + child_folders[0] + 'fills_and_bmodes.pkl'
     csv_file_names = ['fill_csvs/fill_%d.csv']
-    h5_file_names = ['heatloads_fill_h5s/heatloads_all_fill_%i.h5']
     filling_pattern_csv = base_folder + child_folders[0] + 'injection_scheme_2015.csv'
 else:
     raise ValueError('Invalid year!')
@@ -72,29 +79,8 @@ arc_correction_factor_list = hl.arc_average_correction_factors()
 arcs_variable_list = hl.average_arcs_variable_list()
 first_correct_filln = 4474 # from 016_
 def correct_hl(heatloads):
-    for factor,arc_variable in zip(arc_correction_factor_list, arcs_variable_list):
+    for factor, arc_variable in zip(arc_correction_factor_list, arcs_variable_list):
         heatloads.timber_variables[arc_variable].values *= factor
-
-# Proper keys for the output dictionary
-re_arc = re.compile('(S\d\d)_QBS_AVG_ARC.POSST')
-re_quad = re.compile('(Q)RL[^A]\w_0(\d[LR]\d)_QBS\d{3}.POSST')
-re_special = re.compile('QRLAA_(\d{2}\w\d)_QBS\d{3}(_\w\w)?.POSST')
-re_list = [re_arc, re_quad, re_special]
-
-def output_key(input_key, strict=True):
-    for regex in re_list:
-        info = re.search(regex, input_key)
-        if info is not None:
-            out = ''
-            for i in info.groups():
-                if i != None:
-                    out += i
-            return out
-    else:
-        if strict:
-            raise ValueError('No match for %s' % var)
-        else:
-            return input_key
 
 # Other functions
 def add_to_dict(dictionary, value, keys, zero=False):
@@ -121,19 +107,16 @@ def cast_to_na_recursively(dictionary, assure_length=None):
         else:
             log_print('Unexpected type in dictionary for key %s' % key)
 
-def data_integration(timestamps, values):
+def data_integration(timestamps, values, key):
     # Trapezoidal integration
     output = 0.
     nan = np.isnan(values)
-    if np.sum(nan) > 0:
-        log_print('Fill %i: There have been nan values for var %s' % (filln,var))
-        values[nan] = 0.
+    values[nan] = 0.
+    #if np.sum(nan) > 0:
+        #log_print('Fill %i: There have been nan values for var %s' % (filln,key))
     for i in xrange(len(values)-1):
         output += (timestamps[i+1] - timestamps[i])*(values[i] + values[i+1])/2.
     return output
-
-def data_integration_ob(ob, offset=0.):
-    return data_integration(ob.t_stamps, ob.values - offset)
 
 def log_print(*args, **kwargs):
     with open(logfile, 'a') as f:
@@ -141,20 +124,26 @@ def log_print(*args, **kwargs):
     print(*args, **kwargs)
 log_print('%s' % time.ctime())
 log_print('Offset is subtracted?: %s' % subtract_offset)
+log_print('Offset is the average of %i seconds before t_inj_proton' % average_offset_seconds)
 
 # Time keys
-time_key_list = ['start_ramp', 'stop_squeeze', 'stable_beams']
+time_key_list = ['t_start_fill', 't_inj_proton', 'start_ramp', 'stop_squeeze', 'stable_beams']
 for ii in xrange(hrs_after_sb):
     time_key_list.append('sb+%i_hrs' % (ii+1))
 
-# Heat load groups
-groups_dict = hl.groups_dict()
-all_heat_load_vars = []
-for group, varlist in groups_dict.iteritems():
-    all_heat_load_vars += varlist
-hl_var_dict = {}
-for var in all_heat_load_vars:
-    hl_var_dict[var] = {'key': output_key(var, strict=True)}
+# Time stamps
+def get_time(kk):
+    if kk == 0:
+        tt = t_start_fill
+    elif kk == 1:
+        tt = t_start_injphys
+    elif kk == 2:
+        tt = t_start_ramp
+    elif kk == 3:
+        tt = t_stop_squeeze
+    else:
+        tt = t_stable_beams + (kk-4)*3600
+    return tt
 
 # Filling numbers
 with open(fills_bmodes_file, 'r') as f:
@@ -167,23 +156,27 @@ sr_calc = hls.HeatLoadCalculatorSynchrotronRadiationLHCArc()
 
 # Main loop
 output_dict = {}
+
+if args.fills:
+    fills_0 = sorted([int(x) for x in args.fills])
 for filln in fills_0:
     process_fill = True
 
     # Check if this fill reached stable beams
+    t_start_fill = fills_and_bmodes[filln]['t_startfill']
+    t_start_injphys = fills_and_bmodes[filln]['t_start_INJPHYS']
     t_stable_beams = fills_and_bmodes[filln]['t_start_STABLE']
-    t_begin_inj = fills_and_bmodes[filln]['t_start_INJPROT']
     if t_stable_beams == -1:
         log_print('Fill %i did not reach stable beams.' % filln)
         process_fill = False
-    elif t_begin_inj == -1:
+    elif t_start_injphys == -1:
         log_print('Warning: Offset for fill %i could not be calculated as t_start_INJPROT is not in the fills and bmodes file!' % filln)
         process_fill = False
 
     # Check if all files exist and store their paths
     if process_fill:
         this_files = []
-        for f in csv_file_names+h5_file_names:
+        for f in csv_file_names:
             f = f % filln
             for child in child_folders:
                 path = base_folder + child + f
@@ -214,6 +207,7 @@ for filln in fills_0:
     if process_fill:
         n_tries = 0
         while n_tries < 5:
+            n_tries += 1
             try:
                 qbs_ob = qf.compute_qbs_fill(filln)
                 break
@@ -221,63 +215,28 @@ for filln in fills_0:
                 log_print('Fill %i: No recomputed data: %s!' % (filln,e))
                 # Suspicious fails of read attempts -> try once more
                 time.sleep(5)
-                n_tries += 1
         else:
             process_fill = False
             log_print('Fill %i: Recomputed data read attempt failed!' % filln)
 
         if process_fill:
-            fill_dict.update(qf.get_fill_dict(qbs_ob))
             lhc_hl_dict = qf.lhc_arcs(qbs_ob)
+            arc_averages = qf.compute_qbs_arc_avg(qbs_ob)
 
     # Main part - obtain and store the variables of interest
     if process_fill:
         log_print('Fill %i is being processed.' % filln)
 
         ## Allocate objects that are used later
-        en_ob = energy(fill_dict, beam=1)
-        heatloads = SetOfHomogeneousNumericVariables(all_heat_load_vars, fill_dict)
-        bct_bx = {}
-        blength_bx = {}
-        fbct_bx = {}
-        for beam_n in (1,2):
-            bct_bx[beam_n] = BCT(fill_dict, beam=beam_n)
-            blength_bx[beam_n] = blength(fill_dict, beam=beam_n)
-            fbct_bx[beam_n] = FBCT(fill_dict, beam=beam_n)
-
-        # Correct Arc Averages
-        if filln < first_correct_filln:
-            correct_hl(heatloads)
-
-        # Heat load offset
-        if subtract_offset:
-            this_subtract_offset = True
-            offset_dict = {}
-            if t_begin_inj == -1:
-                this_subtract_offset = False
-            else:
-                for var in all_heat_load_vars:
-                    hl_ob = heatloads.timber_variables[var]
-                    offset_dict[var] = hl_ob.calc_avg(t_begin_inj, t_begin_inj+600)
-        else:
-            this_subtract_offset = False
-
+        en_ob      = energy(fill_dict, beam=1)
+        bct_bx     = {beam_n: BCT(fill_dict, beam=beam_n)     for beam_n in (1,2)}
+        blength_bx = {beam_n: blength(fill_dict, beam=beam_n) for beam_n in (1,2)}
+        fbct_bx    = {beam_n: FBCT(fill_dict, beam=beam_n)    for beam_n in (1,2)}
 
         ## Populate output dict
 
         # Fill Number
         add_to_dict(output_dict, filln, ['filln'])
-
-        # Integrated heat load
-        for var in all_heat_load_vars:
-            hl_ob = heatloads.timber_variables[var]
-            if this_subtract_offset:
-                offset = offset_dict[var]
-            else:
-                offset = 0.
-            summed_hl = data_integration_ob(hl_ob, offset)
-            key = hl_var_dict[var]['key']
-            add_to_dict(output_dict, summed_hl, ['integrated_hl', key])
 
         # Filling pattern and bpi
         pattern = filling_pattern_ob.nearest_older_sample(t_stable_beams)[0]
@@ -297,20 +256,12 @@ for filln in fills_0:
         t_start_ramp = fills_and_bmodes[filln]['t_start_RAMP']
         t_stop_squeeze = fills_and_bmodes[filln]['t_stop_SQUEEZE']
         end_time = fills_and_bmodes[filln]['t_endfill']
+
         for kk, time_key in enumerate(time_key_list):
-            if kk == 0:
-                tt = t_start_ramp
-            elif kk == 1:
-                tt = t_stop_squeeze
-            else:
-                tt = t_stable_beams + (kk-2)*3600
+            tt = get_time(kk)
             # zero controls if calculations for output are performed.
             # If zero is True, then only 0s are stored in the output_dict
-            if tt > end_time:
-                zero = True
-            else:
-                zero = False
-
+            zero = tt > end_time
             this_add_to_dict = lambda x, keys: add_to_dict(output_dict, x, [time_key]+keys, zero=zero)
 
             # t_stamps
@@ -384,28 +335,17 @@ for filln in fills_0:
             this_add_to_dict(tot_imp+tot_sr, ['heat_load', 'total_model'])
 
             # Heat loads
-            for var in all_heat_load_vars:
-                hl_ob = heatloads.timber_variables[var]
-                if this_subtract_offset:
-                    offset = offset_dict[var]
-                else:
-                    offset = 0
-                if not zero:
-                    try:
-                        hl = hl_ob.nearest_older_sample(tt) - offset
-                    except:
-                        log_print('Fill %i: Warning: No hl data for %s' % (filln,var))
-                        hl = 0
-                else:
-                    hl = 0
-                key = hl_var_dict[var]['key']
-                this_add_to_dict(hl, ['heat_load', key])
-                this_add_to_dict(hl, ['hl_offset', key])
-            # Recalculated cells
-            for arc, atd in lhc_hl_dict.iteritems():
-                values = atd.nearest_older_sample(tt)
-                for cell_ctr, cell in enumerate(atd.variables):
-                    this_add_to_dict(values[cell_ctr], ['heat_load_re', arc, cell])
+            for obj, main_key in zip([qbs_ob, arc_averages], ['all_cells', 'arc_averages']):
+                index = np.argmin(np.abs(obj.timestamps - tt))
+                mask_offset = np.logical_and(obj.timestamps < t_start_injphys, obj.timestamps > t_start_injphys - average_offset_seconds)
+                for key, arr in obj.dictionary.iteritems():
+                    offset = np.mean(arr[mask_offset])
+                    hl = arr[index] - offset
+                    integrated_hl = data_integration(obj.timestamps, arr-offset, key)
+                    this_add_to_dict(hl, ['heat_load', main_key, key])
+                    this_add_to_dict(offset, ['hl_subtracted_offset', main_key, key])
+                    this_add_to_dict(offset, ['hl_integrated', main_key, key])
+
 
 n_fills = len(output_dict['filln'])
 cast_to_na_recursively(output_dict, assure_length=n_fills)
@@ -413,6 +353,11 @@ cast_to_na_recursively(output_dict, assure_length=n_fills)
 # Dump this dict
 with open(pkl_file_name, 'w') as f:
     cPickle.dump(output_dict, f, protocol=-1)
-os.remove(latest_pkl)
-os.symlink(os.path.basename(pkl_file_name), latest_pkl)
-log_print('\nSuccess\n')
+
+if not args.o:
+    os.remove(latest_pkl)
+    os.symlink(os.path.basename(pkl_file_name), latest_pkl)
+
+log_print('\nSuccess')
+log_print('Saved to %s\n' % pkl_file_name)
+
